# specs/004-swagger-docs/contracts/openapi-schema-example.yaml
# OpenAPI 3.0 Schema Example
# This is an example of what the generated OpenAPI schema will look like
# The actual schema will be auto-generated by @fastify/swagger from route definitions

openapi: 3.0.0
info:
  title: License API
  description: Backend API for the License Management System
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Products
    description: Product management endpoints
  - name: Plans
    description: Plan management endpoints
  - name: Licenses
    description: License management endpoints
  - name: Validation
    description: License validation endpoints (public)
  - name: Consumption
    description: Usage consumption endpoints (public)
  - name: Logs
    description: Event log endpoints
  - name: Health
    description: Health check endpoint (public)

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns server health status
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /api/products:
    get:
      tags:
        - Products
      summary: List all products
      description: Retrieve all products in the system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Products
      summary: Create a new product
      description: Create a new product with name, code, and optional description
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            examples:
              example1:
                value:
                  name: "My Product"
                  code: "PROD001"
                  description: "Product description"
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/products/{id}:
    delete:
      tags:
        - Products
      summary: Delete a product
      description: Delete a product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/validate:
    post:
      tags:
        - Validation
      summary: Validate a license
      description: Validate a license key and register device activation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'

  /api/consume:
    post:
      tags:
        - Consumption
      summary: Consume usage from license
      description: Consume a specified amount of usage from a usage-based license
      parameters:
        - name: product-id
          in: header
          required: true
          schema:
            type: string
          description: Product ID for validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRequest'
      responses:
        '200':
          description: Usage consumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token authentication

  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - code
      properties:
        id:
          type: string
          description: Unique product identifier
        name:
          type: string
          description: Product name
        code:
          type: string
          description: Product code (unique identifier)
        description:
          type: string
          description: Product description
          nullable: true

    CreateProductRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          description: Product name
        code:
          type: string
          description: Product code (unique identifier)
        description:
          type: string
          description: Product description
          nullable: true

    Plan:
      type: object
      required:
        - id
        - productId
        - name
        - features
        - price
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        defaultUsageCount:
          type: integer
          description: For usage-based plans only
          minimum: 1
        durationDays:
          type: integer
          description: For time-based plans only
          minimum: 1
        deviceLimit:
          type: integer
          description: For time-based plans only
          minimum: 0
        features:
          type: array
          items:
            type: string
        price:
          type: number
          minimum: 0

    License:
      type: object
      required:
        - id
        - key
        - productId
        - planId
        - userEmail
        - status
        - issuedAt
        - expiresAt
        - activations
      properties:
        id:
          type: string
        key:
          type: string
        productId:
          type: string
        planId:
          type: string
        userEmail:
          type: string
          format: email
        status:
          type: string
          enum:
            - ACTIVE
            - EXPIRED
            - CANCELLED
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        activations:
          type: array
          items:
            $ref: '#/components/schemas/Activation'
        currentUsageCount:
          type: integer
          description: For usage-based licenses only
        totalUsageCount:
          type: integer
          description: For usage-based licenses only

    Activation:
      type: object
      required:
        - deviceId
        - activatedAt
        - lastUsedAt
      properties:
        deviceId:
          type: string
        activatedAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time

    ValidationRequest:
      type: object
      required:
        - key
        - deviceId
      properties:
        key:
          type: string
          description: License key
        deviceId:
          type: string
          description: Device identifier

    ValidationResponse:
      type: object
      required:
        - valid
        - message
      properties:
        valid:
          type: boolean
        message:
          type: string
        license:
          type: object
          properties:
            status:
              type: string
              enum:
                - ACTIVE
                - EXPIRED
                - CANCELLED
            productName:
              type: string
            planName:
              type: string
            expiresAt:
              type: string
              format: date-time
            features:
              type: array
              items:
                type: string
            daysRemaining:
              type: integer

    ConsumptionRequest:
      type: object
      required:
        - key
        - amount
      properties:
        key:
          type: string
          description: License key
        amount:
          type: integer
          description: Amount to consume
          minimum: 1

    ConsumptionResponse:
      type: object
      required:
        - success
        - remaining
        - message
      properties:
        success:
          type: boolean
        remaining:
          type: integer
          description: Remaining usage count
        message:
          type: string

    LogEvent:
      type: object
      required:
        - id
        - timestamp
        - type
        - details
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        type:
          type: string
          enum:
            - ISSUE
            - VALIDATE
            - RENEW
            - CANCEL
            - EXPIRE
            - ERROR
            - CONSUME
            - TOPUP
        details:
          type: string
        relatedId:
          type: string
          nullable: true

    HealthStatus:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          example: "ok"
        timestamp:
          type: string
          format: date-time

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        hint:
          type: string
          description: Additional hint for debugging

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            code: "VALIDATION_ERROR"

    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication failed"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"